+++ crates/Cargo.toml
@@ -5,4 +5,5 @@ members = [
     "extracted_fzy",
     "fuzzy_filter",
     "icon",
+    "printer"
 ]
